While we're here, I'll add a use statement for MinkContext and make it extend that. I'll explain that in a minute.
One last bit of setup: at the root of your project create a behat.yml file. I'll paste in some content to get us started.

When we run Behat it will looks for a behat.yml file and this tells it:

Yo! Our application lives at localhost, so look for it there.