Our library is done downloading and we just need to activate it in our behat.yml with the line:
 7 lines  behat.yml
default:
    extensions:
      Behat\MinkExtension:
... lines 4 - 5
        selenium2: ~
This gives me the option to use goutte to run the test using curl requests or Selenium to have things run in a browser. By default, this will just select goutte. So how do we make it use Selenium? I'm so glad you asked!

Above the scenario that you want to run in Selenium add @javascript:

 12 lines  features/search.feature
... lines 1 - 5
  @javascript
  Scenario: Search for a word that exists
... lines 8 - 12
And that's it. Go back to the terminal and let's rerun this test. It actually opens the browser, it's quick but you can see it clicking around to complete the scenario. Cool!

 Tip
FireFox is buggy with the new Selenium 3 server that's why it's preferable to use Google Chrome. You can explicitly specify the browser in the behat.yml config file:

# behat.yml
default:
  extensions:
    Behat\MinkExtension:
      browser_name: chrome
We write human readable instructions and they turn into functional tests, and this just barely scratches the surface of how this will change your development. Let's keep going and figure out what's really going on here.