Writing Features
Our application is already partially built, but let's pretend that it isn't and we're just in the planning stages. First we know we'll need an authentication feature. So let's go into the features directory and create a new authentication.feature file. Each feature will have it's own file in here.

We always start with the same header Feature and a short description of it. Here we'll just say 'Authentication'. The next three lines are very important. Our next line is always 'In order to' followed by the business value. In the case of the Raptor Store, since you need to login to see the product area, I would say that the business value is 'to gain access to the management area'.

Next is 'As a' and you say who is going to benefit from this feature, in our case it would be an admin user. And the third line is, 'I need to be able to' followed by a short description of what the user would actually be able to do with this feature. In our store that is 'login and logout'.

The most important parts are the first two lines, the business value and the user - or role - that's going to benefit from this value. If either of these are difficult to define then maybe your feature isn't actually valuable. Or maybe none is benefiting from it. Perhaps move onto a different task.