BDD with a New Scenario
Let's create a new scenario for functionality that doesn't actually exist yet: adding a new product. Time to plan out this scenario!

We don't need to add any products in the database before hand so we just start with. We're going to want a button on this page that says 'New Product'. So let's add clicking that to the process of creating a new product. Clicking this will take us to another page with a form containing fields for name, price and description. To keep our When going to include the action of filling out this form we'll use And.

There is a secret reason of why I'm using language like 'I fill in' after the And, which we'll cover soon. Lastly, in our process I'll expect to hit a save button to finish creating my new product. So let's add another line to our scenario. At this point my product should be saved and I will be redirected to another page with a success message to prove that this worked. This is where we'll say.

Isn't that a nice way to plan out your user flow? When we do finally code this, we'll know exactly how everything should work.