Ok, the basics of Gherkin, writing features and scenarios, are now behind us. Now, how does this 'Behat' thing fit in?
Imagine we've gone back in time 25 years and Linus Torvalds, the Yoda of Linux, comes to us and says:

I would love your help in building the ls command.
Yep, the ls command right here in the terminal. Since you're an awesome dev, you reply:

I would be happy to help you Linus, and I'll use Gherkin to describe the feature and the scenarios for the ls command. I'm really into Behavior Driven Development.
Writing Scenario #1

Create a new ls.feature file. I'll save some time by copying the feature description in since we've already covered this pretty well. On to the scenario!

And now we'll go through our Given, When, and Then lines. Since we need to list two files we'll need to create those first.

The user action would be actually running the ls command.

Finally, we'll actually test that the 'john' and 'hammond' files both appear in the output.

Writing this scenario has two purposes. First, it lets us plan how our feature should behave. And that's what we've been talking about. Second, we want the scenario to be executed as a test to prove whether or not we have successfully created this behavior. To do that, we'll run ./vendor/bin/behat in our terminal and point it at features/ls.feature and let's see what happens!