The Essence of Behat: Matching Scenario lines with Functions
Ahh, so it says our scenario was undefined and something about our FeatureContext having missing steps. And it even gives us some PHP code. Copy these three functions, open up the FeatureContext class that we generated and paste them there.

So what does Behat really do? Simply, it reads each line of our scenario, looks for a matching function inside of FeatureContext and calls it. In this case, it'll read 'There is a file named 'john', find that it matches this annotation here and then execute its function.

What's really cool is that because we surrounded john with quotes, it recognized that as a wildcard. In the annotation, we have :arg1 which means it matches anything surrounded in quotes or a number.

Filling in the Definitions/Functions

Our job is just to make these functions do what they say they will do. I'll change this to :filename and update the argument to , just because that's more descriptive. In this function, how do we create a file? How about we use touch({dollar_sign}filename). For iRun() update the arg1's to command. There are lots of ways to run commands, but we'll use shell_exec({dollar_sign}command)

Sharing Data inside your Scenario (between 'Steps')

Lastly, in iShouldSeeInTheOutput(), update the arg1 to string. And now we're stuck... we don't have the return value from shell_exec() above.