Failing!
Now in iShouldSeeInTheOutput() method if (strpos({dollar_sign}this->output, {dollar_sign}string) === false)) then we have a problem and we want this step to fail.

How do you fail in Behat? By throwing an exception: throw new \Exception() and print a really nice message here of 'Did not see '%s' in the output '%s'. Finish that line up with {dollar_sign}string, {dollar_sign}this->output.

Ok let's give this a try!

Re-run our last Behat command in the terminal, and this time it's green! And if you run the ls command here you can see the 'john' and 'hammond' files listed.