What are Definitions and Steps?
Back to our scenario. Get out some pen and paper: we need to review some terminology! Every line in here is called a 'step':
12 lines  features/ls.feature
... lines 1 - 6
  Given I have a file named 'john'
  And I have a file named 'hammond'
  When I run 'ls'
  Then I should see 'john' in the output
  And I should see 'hammond' in the output

And the function that a step connects to is called a 'step definition':
53 lines  features/bootstrap/FeatureContext.php
... lines 1 - 11
class FeatureContext extends MinkContext implements Context, SnippetAcceptingContext {
... lines 14 - 26
  /**
   * @Given I have a file named :filename
   */
  public function iHaveAFileNamed({dollar_sign}filename)
... lines 31 - 34
  /**
   * @When I run :command
   */
  public function iRun({dollar_sign}command)
... lines 39 - 42
  /**
   * @Then I should see :string in the output
   */
  public function iShouldSeeInTheOutput({dollar_sign}string)
... lines 47 - 51
}

This is important in helping you understand Behat's documentation.

Oh, and the 4 feature lines above: those aren't parsed by Behat:
12 lines  features/ls.feature
Feature: ls
  In order to see the directory structure
  As a UNIX user
  I need to be able to list the current directory's contents

We only write those to go through the exercise of thinking about our business value.

Now, you see in our test it says '5 steps (5 passed)', which means that each step could fail. The rule is really simple, if the definition function for a step throws an exception, it's a failure. If there's no exception it passes.

Head back to the step that looks for the 'hammond' file and add the number 2 to the end of the file name so it fails. Running the scenario in our terminal now shows us 4 steps passed and 1 failed.