We need to run some code after every scenario. Behat has a system called 'hooks' where you can make a function inside of your context and tell Behat to call it before or after your scenario, entire test suite or individual steps.
Create a new public function inside called public function moveOutOfTestDir(). This will be our cleanup function. Use chdir('..'); to go up one directory. Then, if the test/ directory exists - which it should - then we'll run a command to remove that.
To get Behat to actually call this after every scenario, add an @AfterScenario annotation above the method.

That's it!

Let's give this a try:

vendor/bin/behat features/ls.feature
The first time we run this we still get the warning since our clean up function hasn't been called yet. But when we run it again, the warnings are gone! And if we run ls, we see that there is no test directory.