I installed PHPUnit because it has really nice assert functions that we can get a hold of. To get access to them we just need to add a require statement in our FeatureContext.php file, require_once then count up a couple of directories and find vendor/phpunit/phpunit/src/Framework/Assert/Functions.php.
Requiring this file gives you access to all of PHPUnit's assert functions as flat functions. Down in the iShouldSeeInTheOutput() method, use assertContains(), give it the needle which is {dollar_sign}string and the haystack which is {dollar_sign}this->output. Finally, add our helpful message which I'll just cut and paste. Remove the rest of the original if statement.

Run the test again!

vendor/bin/behat features/ls.feature
Beautiful, it looks just like it did before.