Forget about Behat for a few minutes - I want to talk about a totally different library. Imagine if there were a simple PHP library that let you take command of a browser: surf to pages, find links with CSS, click those links, fill out forms and anything else you can dream up. It exists! And it's called Mink.
This is such a cool library that it deserves some direct attention. Start by creating a mink.php file right at the root of your project. And require composer's autoload file.

We'll use Mink all by itself, outside of Symfony, Behat and everything else so we can focus on just how it works.

Important Object 1: The Driver

As awesome as Mink is, it only has four important objects. The first is the driver. Create a new  variable and set it to new GoutteDriver()

Now, ignore it: I'll come back and explain important object number 1 in a little while.

Important Object 2: The Session

Let's move on to important object #2, and the first that we really care about: the session. Add {dollar_sign}session = new Session() and pass it the {dollar_sign}driver as an argument.

Think of the session like a browser tab: anything you can do in a tab, you can do in a session. And actually, that isn't very much. You can visit URLs, refresh, go backwards, go forwards and that's about it. Let's use it to visit a very awesome and absurdly-designed site 'jurassicpark.wikia.com'. After that we'll just print out a few things about the page like the status code, and the current URL

To execute this, head over to the terminal and run:

php mink.php
Look at that: it's printing out 200 and a slightly different URL than we put in our code. That makes sense: when you go to the site in a browser, it redirects to the URL we see in our terminal. Mink is emulating a real browser by following redirects. But in reality, so far, it's making invisible cURL requests: it's not using a real browser.