Important Object 3: The Page (DocumentElement)
The third important object is called the page. Grab it by saying {dollar_sign}page = {dollar_sign}session->getPage()

I want you to think of this as the JQuery object or the DOM. Anything you can do with JQuery - like select elements, click links and fill out fields - you can do with the page. Less impressively, it also knows the HTML of whatever page we're currently on.

 Tip
If you like to dig into the source code, the page is an instance of DocumentElement.

Let's use it to print out this first bit of text on the page with var_dump(substr({dollar_sign}page->getText()), 0, 75)

Run that again in the terminal.

php mink.php
Now we see the thrilling text of: 'Park Pedia - Jurassic Park, Dinosaurs, Stephen Spielberg...'. There's some weird a:lang code stuff on the end. Open up the source on the page.

The getText() method returns anything other than the HTML tags themselves. The first part comes from the title tag and then it grabs some other stuff from the style tag, which is technically text.

But what we really want to do is find individual elements so we can click links and fill out fields. Let's talk about that next.