Tip This elemens was an h1 tag (when we recorded the video) - now the site has an h2 tag. The code below was updated.
Back on the weird Wiki page, run an inspect element on the navigation. There's a hidden h2 tag inside of the WikiHeader and WikiNav elements. Let's try to find this and print out its text.

Finding Elements by CSS

To do that use the find() function: pass it css as the first argument and then use your css selector: .WikiHeader .WikiNav h2

Important Object 4: An Element (NodeElement)

Surprise! This is actually the fourth - and final - important object in Mink. You start with the page, but as soon as you find a single element, you now have a NodeElement. What's cool is that this new object has all the same methods as the page, plus a bunch of extras that apply to individual elements.

Let's dump the {dollar_sign}header->getText().

And re-run the mink file. Now it prints 'Jurassic Park Wiki Navigation' - so finding by CSS is working.