Find via the Amazing Named Selector
In addition to CSS, there's one more important way to find things using Mink: it's called the named selector. I'm going to paste in some code here: please do not write this -- it's ugly code -- I'll show you a better way.

Instead of passing css to find(), this passes named along with an array that says we're looking for a 'link' whose text is 'Wiki Activity'. The named selector is all about finding an element by its visible text. To see if this is working let's var_dump({dollar_sign}linkEl->getAttribute('href'))

That should come back as the URL to the activity section. Try it out.

php mink.php
It works! The named selector is hugely important because it lets us find elements by their natural text, instead of technical CSS classes. In this case, we're using the text of the anchor tag. But the named selector also looks for matches on the title attribute, on the alt attribute of an image inside of a link and several other things. It finds elements by using anything that a user or a screen reader thinks of as the 'text' of an element.