Behat parses scenarios and Mink is really good at browsing the web. If we combine their powers, we could start having steps that look a lot like what we have in search.feature.
 11 lines  features/search.feature
... lines 1 - 5
Scenario: Search for a word that exists
  Given I am on '/'
  When I fill in 'searchTerm' with 'Samsung'
  And I press 'search_submit'
  Then I should see 'Samsung Galaxy S II'
For the Background step, we now know we could create a matching definition in FeatureContext and easily use Mink's session object to actually go to that URL. But earlier when we ran this scenario, it worked... so there must already be something tie'ing Behat and Mink together.

Let's see what's happening.

Free Behat Steps

First, in FeatureContext I had you extend MinkContext. Remove that now.

When we run Behat, it needs to know all of the step definition language that's available. You can see that list by passing a -dl to the Behat command:

php vendor/bin/behat -dl
This shows the four ls definitions we built. So, Behat opens the FeatureContext class, parses out all of the @Given, @When and @Then annotations, and prints a final list here for our enjoyment.

When we add more step definitions, this list grows. And if we use something that isn't here yet, Behat very politely prints out the function for us in the terminal.