Add a second scenario to search: searching for a product that is not found. I'm lazy, so I'll copy the first scenario.
The Given is the same in both, so we can move that to a Background.

It is ok not to have any Given's in the scenario and start directly with When. I'll change the search term to 'XBox' - none of those in the dino store. This is really going to bum out any dino's that are looking to play Ghost Recon.

When there are no matches - the site says 'No Products Found'. Cool, let's plug that into our Then line.

Run this:

./vendor/bin/behat
We didn't use any new language, so everything runs and passes!