In search.feature:
 18 lines  features/search.feature
... lines 1 - 9
  When I fill in 'searchTerm' with '<term>'
  And I press 'search_submit'
... lines 12 - 18
This searchTerm is the name attribute of the search box. And search_submit is the id of its submit button. Well, listen up y'all, I'm about to tell you one of the most important things about working with Behat: Almost every built-in definitions finds elements using the 'named' selector, not CSS.

Cardinal Rule: Avoid CSS in your Scenarios

For example, look at the definition for

I fill in 'field' with 'value'
To use this, you should pass the label text to 'field', not the id, name attribute or CSS selector. Clicking a link is the same. That's done with the

I follow 'link'
Where the link must be the text of the link. If you pass a CSS selector, it's not going to work. If I changed search_submit to be a CSS selector, it'll fail. Believe me, I've tried it a bunch of times.

Got it? Ok: in reality, the named selector lets you cheat a little bit. In addition to the true 'text' of a field, it also searches for the name attribute and the id attribute. That's why our scenario works.

But please please - don't use the name or id. In fact, The cardinal rule in Behat is that you should never use CSS selectors or other technical things in your scenario. Why? Because the person who is benefiting from the feature is a web user, and we're writing this from their point of view. A web user doesn't understand what searchTerm or search_submit means. That makes your scenario less useful: it's technical jargon instead of behavior descriptions.

So why did we cheat? Well, the search field doesn't have a label and the button doesn't have any text. I can't use the named selector to find these, unless I cheat.