Custom Mink Definition
Whenever you want to cheat or can only find something via CSS, there's a simple solution: use new language and create a custom definition. Change the first line.
If I can't target it with real text, I'll just use some natural language. PhpStorm highlights the line because we don't have a definition function matching this text. For the second problem line, use
 18 lines  features/search.feature
... lines 1 - 10
   And I press the search button
... lines 12 - 18
You know the drill: it's time to run the scenario. It prints out the two functions we need to fill in