Getting the Mink Session
Filling these in shouldn't be hard: we're pretty good with Mink. But, how can we access the Mink Session? There's a couple ways to get it, but the easiest is to make FeatureContext extend RawMinkContext.

This gives us access to a bunch of functions: the most important being getSession() and another called visitPath() that we'll use later.

On the first method, change arg1 to term. Once you're inside of FeatureContext it's totally OK to use CSS selectors to get your work done. Back in the browser, inspect the search box element. It doesn't have an id but it does have a name attribute - let's find it by that. Start with {dollar_sign}searchBox = {dollar_sign}this->getSession()->getPage(). Then, to drill down via CSS, add ->find('css', '[name='searchTerm']');. I'm going to add an assertNotNull() in case the search box isn't found for some reason. Fill that in with {dollar_sign}searchBox, 'The search box was not found'

Now that we have the individual element, we can take action on it with one of the cool functions that come with being an individual element, like attachFile, blur, check, click and doubleClick. One of them is setValue() that works for field. Set the value to {dollar_sign}term.

This is a perfect step definition: find an element and do something with it.

To press the search button, we can do the exact same thing. {dollar_sign}button = {dollar_sign}this->getSession()->getPage()->find('css', '#search_submit');. And assertNotNull(, 'The search button could not be found'). It's always a good idea to code defensively. This time, use the press() method.

We're ready to run the scenario again. It passes!

That was more work, but it's a better solution. With no CSS inside of our scenarios, they're less dependent on the markup on our site and this is a heck of a lot easier to understand than before with the cryptic name and ids.