Create a getPage() Shortcut
To save time in the future, create a private function getPage() and return {dollar_sign}this->getSession()->getPage().

I'll put a little PHPDoc above this so next month we'll remember what this is.

Now we can shorten both definition functions a bit with {dollar_sign}this->getPage()

Test the final scenarios out. Perfect! Now we have access to Mink inside of FeatureContext and we know that including CSS inside of scenarios is not the best way to make friends.