Multiple Suites
But we can go further. in behat.yml, check out the suites key. Currently, we have one 'suite' called default.

But you could have many. What's a suite? It's a combination of a set of feature files and the contexts that should be used for them. Think about it: the ls.feature is the only feature that needs CommandLineProcessContext. And every other feature only needs FeatureContext and MinkContext. This is the perfect use-case for a second suite that I'm going to call commands. In this case, only add the CommandLineProcessContext.

Remove that from the default suite.

When you execute Behat, it uses the default suite unless you tell it which one to use with the --suite option. Try it with --suite=commands and then run ls.feature:

./vendor/bin/behat --suite=commands features/ls.feature
Or you can use the -dl option to see only the definition lists associated with the contexts in that suite:

./vendor/bin/behat --suite=commands features/ls.feature -dl
Without --suite, we see definitions for the default suite:

./vendor/bin/behat -dl