Open up authentication.feature, because, you can't do much until you login. Let's add a scenario.
We remember from running behat -dl:

./vendor/bin/behat -dl
That we have a lot of built in language already. Let's save some effort and describe the login process using these.

We want to start on the homepage. PhpStorm tries to help by auto-completing this step, but when I hit tab, it prints it with extra stuff. When you use the -dl option, the 'Given I am on' ends with a bunch of crazy regex. Anytime you see regex like this, it's just forming a wildcard: something where you can pass any value, surrounded by quotes.

Oh, and one other thing: even though each line starts with Given, When and Then in the definition list, that first word doesn't matter. We could actually say:

Then I am on '/'
It doesn't sound right in English, but technically, it would still run.

Alright! Given I am on '/'. Next, I will click 'Login'. The built-in definition for that is 'I follow'. There's no built-in definition for 'I click' but we'll add one later since that's how most people actually talk.

Remember, these all use the named selector, so we're using 'Login' because that's the text of the link.

On the login page, we need to fill in these two fields. Again, because of the named selector, we'll target these by the labels Username and 'Password'. There are a few definitions for fields, byt the one I like is 'When I fill in field with value'.

Yes I know, that isn't the right password - this won't work yet. It's cool though.

Finally, press the login button with.

Notice that you follow a link but you press a button. Then we need to find something to assert - some sign that this worked. Login on the browser. Hmm, nothing says 'Congratulations, you're in! '... but our login button did change to 'Logout'. Let's look for that with.

Good work team. Let's run this!

./vendor/bin/behat features/web/authentication.feature
Debugging Failed Scenarios

It runs... and fails:

The text 'Logout' was not found anywhere in the text of the current page.
I happen to know this is because the password is wrong. But let's pretend that we didn't know that, and we're staring at this message wondering what the heck is going on.