Eventually, you'll need to insert test data at the beginning of your scenarios. And that's when things get tricky. So, let's learn to do this right.
First: never assume that there is data in the database at the beginning of a scenario. Instead, you should put any data you need there intentionally with a Given.

Go to the top of this scenario and add the necesary data.

I'm inventing this language - it describes something that needs to be setup using natural language. Change the Given line below this to And for readability.

Run it again: it should print out the step definition for this new language.

./vendor/bin/behat features/web/authentication.feature
And it does!