Saving a New User
Inside of thethereIsAnAdminUserWithPassword() step definition, let's go to work! I already have a User entity setup in the project, so we can say {dollar_sign}user = new User(). Then set the username and the 'plainPassword'

I have a Doctrine listener already setup that will encode the password automatically. Which is good: it's well-known that raptors can smell un-encoded passwords...

In this app, to make this an 'admin' user, we need to give the user ROLE_ADMIN.

Now the moment of truth: we need the entity manager. Cool! Grab it with {dollar_sign}em = self::{dollar_sign}container->get('doctrine') (to get the Doctrine service) and then ->getManager()
It's easy from here: {dollar_sign}em->persist({dollar_sign}user); and {dollar_sign}em->flush()

And that should do it!

We already have a user called admin in the database, so let's test this using admin2. Give it a go!

./vendor/bin/behat features/web/authentication.feature
This should not work, since this user isn't in the database yet... unless it's created by the code we just wrote! Brilliant!

This is huge: we're guaranteeing there's an admin2 user by bootstrapping our app and being dangerous with all of our useful services.