An extension in Behat is a plugin. We're already using the MinkExtension. Activate the new plugin in behat.yml: Behat\Symfony2Extension.
And as luck would have it, it doesn't need any configuration. It looks like we still need to wait for it to finish installing in the terminal... there we go!

The most important thing the Symfony2 Extension gives you is, access to Symfony's container... but wait, we already have that? Well, this just makes it easier.