Search for the old self::{dollar_sign}container code. Change it to {dollar_sign}this->getContainer()
You see that PhpStorm all of a sudden auto-completes the methods on the services we fetch because it recognizes this as the container and so knows that this returns the entity manager.

Let's try things again!

./vendor/bin/behat features/web/authentication.feature
Still works! But now with less effort. If you have multiple context classes, you can use the KernelDictionary on all of them to get access to the container.